locate fetch in TodoContainer.jsx and some improvements
